+++ electron/app/createWindow.ts
@@ -59,7 +59,7 @@ const userTempDir = app.getPath('temp');
 const pluginsDir = path.join(userDataDir, 'monoklePlugins');
 const templatesDir = path.join(userDataDir, 'monokleTemplates');
 const templatePacksDir = path.join(userDataDir, 'monokleTemplatePacks');
-const APP_DEPENDENCIES = ['kubectl', 'helm', 'kustomize'];
+const APP_DEPENDENCIES = ['kubectl', 'helm', 'kustomize', 'git'];
 const machineId = machineIdSync();
 
 export const createWindow = (givenPath?: string) => {
@@ -72,6 +72,7 @@ export const createWindow = (givenPath?: string) => {
     title: 'Monokle',
     icon: image,
     webPreferences: {
+      zoomFactor: utilsElectronStore.get('ui.zoomFactor'),
       webSecurity: false,
       contextIsolation: false,
       nodeIntegration: true, // <--- flag