+++ src/components/organisms/PageHeader/PageHeader.tsx
@@ -10,14 +10,21 @@ import {ReloadOutlined} from '@ant-design/icons';
 import newGithubIssueUrl from 'new-github-issue-url';
 
 import {TOOLTIP_DELAY} from '@constants/constants';
-import {NotificationsTooltip} from '@constants/tooltips';
+import {InitializeGitTooltip, NotificationsTooltip} from '@constants/tooltips';
 
 import {K8sResource} from '@models/k8sresource';
 
+import {setRepo} from '@redux/git';
 import {useAppDispatch, useAppSelector} from '@redux/hooks';
+import {updateProjectsGitRepo} from '@redux/reducers/appConfig';
 import {setAutosavingError} from '@redux/reducers/main';
 import {setLayoutSize, toggleNotifications, toggleStartProjectPane} from '@redux/reducers/ui';
 import {activeProjectSelector, isInPreviewModeSelector, kubeConfigContextColorSelector} from '@redux/selectors';
+
+import {Icon} from '@components/atoms';
+import BranchSelect from '@components/molecules/BranchSelect';
+
+import {promiseFromIpcRenderer} from '@utils/promises';
 
 import MonokleKubeshopLogo from '@assets/MonokleLogoDark.svg';
 
@@ -32,6 +39,7 @@ const PageHeader = () => {
   const activeProject = useAppSelector(activeProjectSelector);
   const autosavingError = useAppSelector(state => state.main.autosaving.error);
   const autosavingStatus = useAppSelector(state => state.main.autosaving.status);
+  const hasGitRepo = useAppSelector(state => Boolean(state.git.repo));
   const helmChartMap = useAppSelector(state => state.main.helmChartMap);
   const helmValuesMap = useAppSelector(state => state.main.helmValuesMap);
   const isInPreviewMode = useAppSelector(isInPreviewModeSelector);
@@ -42,11 +50,13 @@ const PageHeader = () => {
   const previewResourceId = useAppSelector(state => state.main.previewResourceId);
   const previewType = useAppSelector(state => state.main.previewType);
   const previewValuesFileId = useAppSelector(state => state.main.previewValuesFileId);
+  const projectRootFolder = useAppSelector(state => state.config.selectedProjectRootFolder);
   const resourceMap = useAppSelector(state => state.main.resourceMap);
 
   let timeoutRef = useRef<any>(null);
 
   const [isHelpMenuOpen, setIsHelpMenuOpen] = useState(false);
+  const [isInitializingGitRepo, setIsInitializingGitRepo] = useState(false);
   const [showAutosaving, setShowAutosaving] = useState(false);
 
   const runningPreviewConfiguration = useAppSelector(state => {
@@ -85,6 +95,22 @@ const PageHeader = () => {
 
     shell.openExternal(url);
   }, [autosavingError]);
+
+  const initGitRepo = async () => {
+    if (!projectRootFolder) {
+      return;
+    }
+
+    setIsInitializingGitRepo(true);
+
+    await promiseFromIpcRenderer('git.initGitRepo', 'git.initGitRepo.result', projectRootFolder);
+
+    promiseFromIpcRenderer('git.fetchGitRepo', 'git.fetchGitRepo.result', projectRootFolder).then(result => {
+      dispatch(setRepo(result));
+      setIsInitializingGitRepo(false);
+      dispatch(updateProjectsGitRepo([{path: projectRootFolder, isGitRepo: true}]));
+    });
+  };
 
   useEffect(() => {
     if (previewResourceId) {
@@ -143,6 +169,23 @@ const PageHeader = () => {
             <>
               <S.Divider type="vertical" />
               <ProjectSelection />
+              {hasGitRepo ? (
+                <S.BranchSelectContainer>
+                  <BranchSelect />
+                </S.BranchSelectContainer>
+              ) : (
+                <Tooltip mouseEnterDelay={TOOLTIP_DELAY} placement="bottomRight" title={InitializeGitTooltip}>
+                  <S.InitButton
+                    icon={<Icon name="git" />}
+                    loading={isInitializingGitRepo}
+                    type="primary"
+                    size="small"
+                    onClick={initGitRepo}
+                  >
+                    Initialize Git
+                  </S.InitButton>
+                </Tooltip>
+              )}
               <CreateProject />
             </>
           )}