+++ src/models/ui.ts
@@ -54,6 +54,7 @@ export type MonacoUiState = {
 export type LeftMenuSelectionType =
   | 'file-explorer'
   | 'helm-pane'
+  | 'git-pane'
   | 'kustomize-pane'
   | 'templates-pane'
   | 'images-pane'
@@ -162,7 +163,7 @@ export type UiState = {
 export type PaneConfiguration = {
   leftPane: number;
   navPane: number;
-  rightPane: number;
+  editPane: number;
   bottomPaneHeight: number;
   recentProjectsPaneWidth: number;
 };