+++ src/redux/git/git.initialState.ts
@@ -1,0 +1,7 @@
+import {GitChangedFile, GitSliceState} from '@models/git';
+
+export const gitInitialState: GitSliceState = {
+  selectedItem: {} as GitChangedFile,
+  changedFiles: [],
+  repo: undefined,
+};