+++ src/redux/store.ts
@@ -6,6 +6,8 @@ import {sectionBlueprintMiddleware} from '@src/navsections/sectionBlueprintMiddl
 
 import * as compareListeners from './compare/listeners';
 import {compareSlice} from './compare/slice';
+import {formSlice} from './forms';
+import {gitSlice} from './git';
 import {combineListeners, listenerMiddleware} from './listeners/base';
 import {alertSlice} from './reducers/alert';
 import {configSlice} from './reducers/appConfig';
@@ -51,6 +53,8 @@ const appReducer = combineReducers({
   terminal: terminalSlice.reducer,
   ui: uiSlice.reducer,
   uiCoach: uiCoachSlice.reducer,
+  git: gitSlice.reducer,
+  form: formSlice.reducer,
 });
 
 const rootReducer: typeof appReducer = (state, action) => {