+++ src/redux/thunks/setRootFolder.ts
@@ -9,6 +9,7 @@ import {
   HelmValuesMapType,
   ResourceMapType,
 } from '@models/appstate';
+import {GitRepo} from '@models/git';
 import {RootState} from '@models/rootstate';
 
 import {SetRootFolderPayload} from '@redux/reducers/main';
@@ -21,6 +22,7 @@ import {processResources} from '@redux/services/resource';
 import {createRejectionWithAlert} from '@redux/thunks/utils';
 
 import {getFileStats} from '@utils/files';
+import {promiseFromIpcRenderer} from '@utils/promises';
 import {OPEN_EXISTING_PROJECT, trackEvent} from '@utils/telemetry';
 
 import {fileService} from '@src/infrastructure/controllers/FileController';
@@ -110,6 +112,22 @@ export const setRootFolder = createAsyncThunk<
     numberOfResources: Object.values(resourceMap).length,
   });
 
+  const isFolderGitRepo = await promiseFromIpcRenderer<boolean>(
+    'git.isFolderGitRepo',
+    'git.isFolderGitRepo.result',
+    rootFolder
+  );
+  const gitRepo = isFolderGitRepo
+    ? await promiseFromIpcRenderer<GitRepo>('git.fetchGitRepo', 'git.fetchGitRepo.result', rootFolder)
+    : undefined;
+
+  const gitChangedFiles = isFolderGitRepo
+    ? await promiseFromIpcRenderer('git.getChangedFiles', 'git.getChangedFiles.result', {
+        localPath: rootFolder,
+        fileMap,
+      })
+    : [];
+
   return {
     projectConfig,
     fileMap,
@@ -120,5 +138,7 @@ export const setRootFolder = createAsyncThunk<
     isScanExcludesUpdated: 'applied',
     isScanIncludesUpdated: 'applied',
     alert: rootFolder ? generatedAlert : undefined,
+    gitChangedFiles,
+    gitRepo,
   };
 });