+++ src/styles/Colors.ts
@@ -24,6 +24,9 @@ enum Colors {
   grey3 = '#262626', // gray, gray 3
   grey1 = '#141414', // gray, gray 1
 
+  coldGrey = '#31393C',
+  warmGrey = '#222222',
+
   // Notifications
   greenOkay = '#09b89d',
   polarGreen = '#6ABE39',
@@ -61,8 +64,9 @@ enum Colors {
   purple8 = '#AB7AE0',
 
   cyan = '#58D1C9',
-  cyan7 = '#33BCB7',
   cyan5 = '#138585',
+  cyan7 = '#33BCB7',
+  cyan8 = '#58D1C9',
   lightSeaGreen = '#13a8a8',
 
   blue6 = '#1890FF', // Daybreak Blue